Adjust root namespace and move files around.
