+++ src/rapidcore.redis.csproj
@@ -1,9 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netstandard1.6</TargetFramework>
-    <RootNamespace>RapidCore</RootNamespace>
+    <RootNamespace>RapidCore.Redis</RootNamespace>
     <Version>0.0.0</Version>
-    <Title>RapidCore</Title>
+    <Title>RapidCore.Redis</Title>
     <Authors>nover</Authors>
     <Description>RapidCore.Redis is redis dependend code for the RapidCore library</Description>
     <PackageTags>dotnet core;core;rapidcore;business;business logic;help;helpers;library;redis;lock;locking;distributed locking;</PackageTags>
