+++ test/functional/functionaltest.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netcoreapp1.1</TargetFramework>
-    <AssemblyName>functionaltest</AssemblyName>
-    <RootNamespace>RapidCore.FunctionalTest</RootNamespace>
+    <AssemblyName>RapidCore.Redis.FunctionalTest</AssemblyName>
+    <RootNamespace>RapidCore.Redis.FunctionalTest</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0" />
