+++ test/unit/Locking/RedisDistributedAppLockTest.cs
@@ -2,11 +2,11 @@ using System;
 using System.Threading.Tasks;
 using FakeItEasy;
 using RapidCore.Locking;
-using RapidCore.Locking.Redis;
+using RapidCore.Redis.Locking;
 using StackExchange.Redis;
 using Xunit;
 
-namespace RapidCore.UnitTests.Locking.Redis
+namespace RapidCore.Redis.UnitTests.Locking
 {
     public class RedisDistributedAppLockTest
     {
